-Refractoring du code et transfert de certaines fonctions vers les classes de service afin de ne pas avoir à les réécrire plus tard.
