+++ src/main/webapp/app/components/cart/service/cart.service.ts
@@ -6,6 +6,7 @@ import { IProductCart } from 'app/entities/product-cart/product-cart.model';
 import { createRequestOption } from '../../../core/request/request-util';
 import { getTotalCartItems, getTotalCartPrice, ICart } from '../../../entities/cart/cart.model';
 import { ProductToCartService } from '../../products/service/product-to-cart.service';
+import { IProduct } from '../../../entities/product/product.model';
 
 export type EntityResponseType = HttpResponse<IProductCart>;
 export type EntityArrayResponseType = HttpResponse<IProductCart[]>;
@@ -50,5 +51,36 @@ export class CartService {
     return this.http.delete(`${this.resourceUrlCart}/productCart/${idProductCart}`, {
       observe: `response`,
     });
+  }
+
+  updateQuantityProduct(item: IProduct, quantity: number): void {
+    quantity > 0 ? this.updateProductCartQuantity(item, quantity) : this.deleteProductCart(item);
+  }
+
+  private updateProductCartQuantity(item: IProduct, quantity: number): void {
+    this.queryQuantityProduct(item.id!, quantity).subscribe(() => {
+      // Reload component
+      const productLine: IProductCart | undefined = this.productToCartService.cart?.lines?.find(line => line.product?.id === item.id);
+      if (productLine != null) {
+        productLine.quantity = quantity;
+        this.calcTotal();
+      }
+    });
+  }
+
+  private deleteProductCart(product: IProduct): void {
+    const lineProduct: IProductCart | undefined = this.productToCartService.getProductCart(product);
+    if (lineProduct?.id != null) {
+      this.queryDeleteProductCart(lineProduct.id).subscribe(() => {
+        // Reload component
+        if (this.productToCartService.cart?.lines != null) {
+          const indexProductCart = this.productToCartService.cart.lines.indexOf(lineProduct);
+          // Splice is a method to delete starting from <index> a given <number of elements>.
+          this.productToCartService.cart.lines.splice(indexProductCart, 1);
+          this.productToCartService.buildCartContentMap();
+          this.calcTotal();
+        }
+      });
+    }
   }
 }
